Initialize the SVH in a separate thread
This is required to meet ROS2-control's timing requirements.
Putting this in `configure` is not sufficient. The ROS2-controllers will
time-out.

The `read()` mechanism will deliver fresh data once the SVH is done with
its initialization. We might need to check the thread safety of
`isConnected()` and `isHomed()` in the library.